Service/app/requirements.txt
fastapi
uvicorn
ultralytics
pillow
numpy
python-multipart